major horse system update

+++ Djoe-Core/Server/Entities/Player/PlayerData.cs
@@ -26,7 +26,7 @@ namespace Server
         public Character SkinPlayer;
         public int Hunger = 100;
         public int Thirst = 100;
-        public float Health = 100;
+        public int Health = 100;
         public VIP Vip;
         public UCoords LastCoord;
